+++ CHANGELOG.md
@@ -7,8 +7,14 @@ This project <em>does not yet</em> adheres to [Semantic Versioning](https://semv
 
 ## [unreleased]
 
+\-
+
+## [0.60.0] - 2026-02-20
+
 ### Added
 
+- Match OTU by TaxonName task
+- DarwinCore summary task 
 - Report per-object duplicate predicate use [#4741]
 - Batch import BibTeX sources option to include Identifier
 - Batch Upload Asserted Distributions now supports a `gazetteer_id` column [#4739]
@@ -17,6 +23,7 @@ This project <em>does not yet</em> adheres to [Semantic Versioning](https://semv
 
 ### Fixed
 
+- Improved scoping of Catalogue of Life Exports
 - Save notice is missing in New type specimen task
 - Fix: Rails now correctly applies time zones for `user_date_start` and `user_date_end` parameters from the `X-Timezone` header [#4721]
 - Classes not being eager loaded in background jobs, causing dwc importer to fail
@@ -27,6 +34,7 @@ This project <em>does not yet</em> adheres to [Semantic Versioning](https://semv
 ### Changed
 
 - Removed unused BibTeX task [#1440]
+- Updated Ruby gems
 
 [#4741]: https://github.com/SpeciesFileGroup/taxonworks/issues/4741
 [#4742]: https://github.com/SpeciesFileGroup/taxonworks/issues/4742
@@ -6090,7 +6098,8 @@ _Special thanks to Tom Klein for his amazing open-source contributions on this r
 - Loosing input page numbers when switching tabs on New Taxon Name task
 
 [#1532]: https://github.com/SpeciesFileGroup/taxonworks/issues/1532
-[unreleased]: https://github.com/SpeciesFileGroup/taxonworks/compare/v0.59.0...development
+[unreleased]: https://github.com/SpeciesFileGroup/taxonworks/compare/v0.60.0...development
+[0.60.0]: https://github.com/SpeciesFileGroup/taxonworks/compare/v0.59.0...v0.60.0
 [0.59.0]: https://github.com/SpeciesFileGroup/taxonworks/compare/v0.58.0...v0.59.0
 [0.58.0]: https://github.com/SpeciesFileGroup/taxonworks/compare/v0.57.0...v0.58.0
 [0.57.0]: https://github.com/SpeciesFileGroup/taxonworks/compare/v0.56.1...v0.57.0